remove prelu fusing logic from js ;
