+++ tfjs-converter/python/tensorflowjs/op_list/basic_math.json
@@ -368,6 +368,30 @@
     ]
   },
   {
+    "tfOpName": "Prelu",
+    "category": "basic_math",
+    "inputs": [
+      {
+        "start": 0,
+        "name": "x",
+        "type": "tensor"
+      },
+      {
+        "start": 0,
+        "name": "alpha",
+        "type": "tensor"
+      }
+    ],
+    "attrs": [
+      {
+        "tfName": "T",
+        "name": "dtype",
+        "type": "dtype",
+        "notSupported": true
+      }
+    ]
+  },
+  {
     "tfOpName": "Relu",
     "category": "basic_math",
     "inputs": [
