+++ tfjs-converter/src/executor/graph_model_test.ts
@@ -29,7 +29,6 @@ const MODEL_URL = `${HOST}/model.json`;
 const RELATIVE_MODEL_URL = '/path/model.pb';
 let model: GraphModel;
 const bias = tfc.tensor1d([1], 'int32');
-const alpha = tfc.scalar(1, 'int32');
 
 const weightsManifest: tfc.io.WeightsManifestEntry[] =
     [{'name': 'Const', 'dtype': 'int32', 'shape': [1]}];
@@ -63,42 +62,6 @@ const SIMPLE_MODEL: tensorflow.IGraphDef = {
     },
     {name: 'Add1', op: 'Add', input: ['Input', 'Const'], attr: {}},
     {name: 'Add', op: 'Add', input: ['Add1', 'Const'], attr: {}}
-  ],
-  versions: {producer: 1.0, minConsumer: 3}
-};
-
-const PRELU_MODEL: tensorflow.IGraphDef = {
-  node: [
-    {
-      name: 'Input',
-      op: 'Placeholder',
-      attr: {
-        dtype: {
-          type: tensorflow.DataType.DT_FLOAT,
-        },
-        shape: {shape: {dim: []}}
-      }
-    },
-    {
-      name: 'Const',
-      op: 'Const',
-      attr: {
-        dtype: {type: tensorflow.DataType.DT_INT32},
-        value: {
-          tensor: {
-            dtype: tensorflow.DataType.DT_INT32,
-            tensorShape: {dim: []},
-          }
-        },
-        index: {i: 0},
-        length: {i: 4}
-      }
-    },
-    {name: 'Relu', op: 'Relu', input: ['Input'], attr: {}},
-    {name: 'Neg', op: 'Neg', input: ['Input'], attr: {}},
-    {name: 'Relu2', op: 'Relu', input: ['Neg'], attr: {}},
-    {name: 'Mul', op: 'Mul', input: ['Const', 'Relu2'], attr: {}},
-    {name: 'Add', op: 'Add', input: ['Relu', 'Mul'], attr: {}}
   ],
   versions: {producer: 1.0, minConsumer: 3}
 };
@@ -145,15 +108,7 @@ const SIMPLE_HTTP_MODEL_LOADER = {
     };
   }
 };
-const PRELU_HTTP_MODEL_LOADER = {
-  load: async () => {
-    return {
-      modelTopology: PRELU_MODEL,
-      weightSpecs: weightsManifest,
-      weightData: alpha.dataSync()
-    };
-  }
-};
+
 const CUSTOM_OP_MODEL: tensorflow.IGraphDef = {
   node: [
     {
@@ -422,29 +377,6 @@ describe('Model', () => {
           {name: 'Add', shape: undefined, dtype: undefined}
         ]);
       });
-    });
-  });
-
-  describe('prelu op model', () => {
-    beforeEach(() => {
-      spyOn(tfc.io, 'getLoadHandlers').and.returnValue([
-        PRELU_HTTP_MODEL_LOADER
-      ]);
-      spyOn(tfc.io, 'browserHTTPRequest')
-          .and.returnValue(PRELU_HTTP_MODEL_LOADER);
-    });
-
-    it('fusePrelu should call model rewrite method', async () => {
-      await model.load();
-      const originalResult = model.predict(tfc.scalar(1)) as tfc.Tensor;
-
-      model.fusePrelu();
-      expect(Object.keys(model.weights)).toContain('Const_neg');
-      expect(model.outputNodes[0]).toEqual('Add_Prelu');
-      const result = model.predict(tfc.scalar(1)) as tfc.Tensor;
-
-      tfc.test_util.expectArraysClose(
-          result.dataSync(), originalResult.dataSync());
     });
   });
