+++ tfjs-converter/src/operations/op_list/basic_math.ts
@@ -212,6 +212,17 @@ export const json: OpMapper[] = [
     ]
   },
   {
+    'tfOpName': 'Prelu',
+    'category': 'basic_math',
+    'inputs': [
+      {'start': 0, 'name': 'x', 'type': 'tensor'},
+      {'start': 0, 'name': 'alpha', 'type': 'tensor'},
+    ],
+    'attrs': [
+      {'tfName': 'T', 'name': 'dtype', 'type': 'dtype', 'notSupported': true}
+    ]
+  },
+  {
     'tfOpName': 'Relu',
     'category': 'basic_math',
     'inputs': [
